+++ CHANGELOG.md
@@ -8,6 +8,17 @@ Features:
   * [#325](https://github.com/bootstrap-ruby/rails-bootstrap-forms/pull/325): Support :prepend and :append for the `select` helper - [@donv](https://github.com/donv).
 
 
+## [Pending Release][] (2016-10-18)
+
+Bugfixes:
+  - Change error class from `has-error` to `has-danger`
+  - Change validation text class from `help-block` to `form-control-feedback`
+  - Change help text class from `help-block` to `form-text text-muted`
+  - Change checkbox classes from `checkbox` to `form-check` and added `form-check-label` to labels
+
+Features:
+  - Update to Bootstrap v4
+
 ## [2.6.0][] (2017-02-03)
 
 Bugfixes: